Move gray scale class out of internal package
