+++ common/core/src/commonMain/kotlin/com/trafi/ui/theme/GrayScale.kt
@@ -1,0 +1,15 @@
+package com.trafi.ui.theme
+
+import com.trafi.ui.theme.internal.Color
+
+class GrayScale(
+    val gray100 : Color,
+    val gray200 : Color,
+    val gray300 : Color,
+    val gray400 : Color,
+    val gray500 : Color,
+    val gray600 : Color,
+    val gray700 : Color,
+    val gray800 : Color,
+    val gray900 : Color,
+)
