+++ common/core/src/commonMain/kotlin/com/trafi/ui/theme/internal/CurrentColorPalette.kt
@@ -1,5 +1,7 @@
 package com.trafi.ui.theme.internal
 
+import com.trafi.ui.theme.GrayScale
+
 class CurrentColorPalette(
     val primary: Color,
     val primaryVariant: Color,
@@ -14,16 +16,4 @@ class CurrentColorPalette(
     val onSurface: Color,
     val onError: Color,
     val grayScale: GrayScale,
-)
-
-class GrayScale(
-    val gray100 : Color,
-    val gray200 : Color,
-    val gray300 : Color,
-    val gray400 : Color,
-    val gray500 : Color,
-    val gray600 : Color,
-    val gray700 : Color,
-    val gray800 : Color,
-    val gray900 : Color,
 )
